Add support in the script to use global supplicant for managing the vdevs
Add two modes of operation (global and normal) to the supplicant script.
In the global mode, the global wpa_supplicant is used to manage all the
vdevs which are operating in station mode. So that it is enough to start a
singe instance of wpa_supplicant and it saves lot of memory.

By default, the scripts operate in the global mode unless it is forced to
normal mode. Before starting the supplicant script, start the global
wpa_supplicant process using the following command.
	$ wpa_supplicant -g /tmp/global -B

If global wpa_supplicant is started with the global control interface in
path other than /tmp/, then the actual control interface path can be
specified while running the scripts using the "-g" argument.

In the normal mode, individual instances of wpa_supplicant is started
to manage every single vdev. To force the supplicant script to operate in
normal mode, just add "-n" to the supplicant script's cmd line arguments.